+++ src/gpu/GridGenerator/io/STLReaderWriter/STLWriter.cpp
@@ -9,7 +9,7 @@
 
 void STLWriter::writeSTL(std::vector<Triangle> &vec, const std::string &name, bool writeBinary)
 {
-    const int size = vec.size();
+    const int size = (int)vec.size();
     *logging::out << logging::Logger::INFO_INTERMEDIATE << "Write " << size << " Triangles to STL : " + name + "\n";
 
     std::ios_base::openmode mode = std::ios::out;
